requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
click==7.1.2
    # via
    #   flask
    #   pip-tools
flask-wtf==0.14.3
    # via -r requirements.in
flask==1.1.2
    # via
    #   -r requirements.in
    #   flask-wtf
greenlet==1.0.0
    # via sqlalchemy
itsdangerous==1.1.0
    # via
    #   flask
    #   flask-wtf
jinja2==2.11.3
    # via flask
markupsafe==1.1.1
    # via
    #   jinja2
    #   wtforms
pep517==0.10.0
    # via pip-tools
pip-tools==6.0.1
    # via -r requirements.in
python-dotenv==0.15.0
    # via -r requirements.in
sqlalchemy==1.4.0
    # via -r requirements.in
toml==0.10.2
    # via pep517
werkzeug==1.0.1
    # via flask
wtforms==2.3.3
    # via flask-wtf

# The following packages are considered to be unsafe in a requirements file:
# pip
